Revert "refactor: migrate @masknet/base (#8969)"
This reverts commit 891da21e7912bbb34db2a616b6bad6141c5f1128.
