--- packages/base/src/SignType/index.ts
@@ -1,5 +1,0 @@
-export enum SignType {
-    Message = 'message',
-    TypedData = 'typedData',
-    Transaction = 'transaction',
-}
