+++ packages/base/src/index.ts
@@ -31,12 +31,4 @@ export {
     isK256PrivateKey,
 } from './WebCrypto/index.js'
 export { CheckedError, OptionalResult, andThenAsync } from './ts-results/index.js'
-
-export * from './utils/index.js'
-export * from './i18n/index.js'
-export * from './Plugin/index.js'
-export * from './NextID/index.js'
-export * from './NextID/type.js'
-export * from './Site/index.js'
-export * from './SignType/index.js'
-export * from './Persona/type.js'
+export { parseURLs } from './utils/index.js'
