+++ packages/base/src/utils/index.ts
@@ -1,4 +1,1 @@
 export * from './parseURLs.js'
-export * from './createLookupTableResolver.js'
-export * from './resolve.js'
-export * from './isPopupPage.js'
