+++ packages/shared-base/package.json
@@ -29,6 +29,7 @@
     "async-call-rpc": "^6.1.5",
     "bignumber.js": "9.1.0",
     "dompurify": "^2.4.1",
+    "i18next-browser-languagedetector": "^7.0.1",
     "idb": "^7.1.1",
     "json-stable-stringify": "^1.0.2",
     "pvtsutils": "^1.3.2",
