+++ packages/shared-base/src/NextID/index.ts
