+++ packages/shared-base/src/NextID/type.ts
