+++ packages/shared-base/src/i18n/register-ns.ts
