+++ packages/shared-base/src/index.ts
@@ -2,13 +2,19 @@ export * from '@masknet/base'
 export * from './Messages/Mask.js'
 export * from './Messages/Shared.js'
 export * from './NameService/index.js'
+export * from './NextID/index.js'
+export * from './NextID/type.js'
+export * from './Persona/type.js'
+export * from './Plugin/index.js'
 export * from './Post/type.js'
 export * from './Pure/index.js'
 export * from './Routes/index.js'
+export * from './Site/index.js'
 export * from './Subscriptions/index.js'
 export * from './collections/index.js'
 export * from './color/index.js'
 export * from './convert/index.js'
+export * from './i18n/index.js'
 export * from './kv-storage/index.js'
 export * from './onDemandWorker/index.js'
 export * from './serializer/index.js'
