+++ packages/shared-base/src/utils/createLookupTableResolver.ts
