+++ packages/web3-helpers/package.json
@@ -11,7 +11,7 @@
   },
   "types": "./dist/index.d.ts",
   "dependencies": {
-    "@masknet/base": "workspace:^",
+    "@masknet/shared-base": "workspace:^",
     "@masknet/web3-shared-base": "workspace:^",
     "@masknet/web3-shared-evm": "workspace:^",
     "@masknet/web3-shared-flow": "workspace:^",
