+++ packages/web3-helpers/src/index.ts
@@ -1,4 +1,4 @@
-import { type NetworkPluginID } from '@masknet/base'
+import { type NetworkPluginID } from '@masknet/shared-base'
 import type {
     Connection,
     ConnectionOptions,
