+++ packages/web3-helpers/tsconfig.json
@@ -7,7 +7,7 @@
   },
   "include": ["./src", "./src/**/*.json"],
   "references": [
-    { "path": "../base" },
+    { "path": "../shared-base" },
     { "path": "../web3-shared/base" },
     { "path": "../web3-shared/evm" },
     { "path": "../web3-shared/solana" },
