+++ packages/web3-shared/base/package.json
@@ -13,9 +13,9 @@
   "dependencies": {
     "@dimensiondev/mask-wallet-core": "0.1.0-20211013082857-eb62e5f",
     "@masknet/public-api": "workspace:^",
+    "@masknet/shared-base": "workspace:^",
     "@servie/events": "^3.0.0",
     "@types/use-subscription": "^1.0.0",
-    "@masknet/base": "workspace:^",
     "bignumber.js": "9.1.0",
     "date-fns": "^2.29.3",
     "react-use": "^17.4.0",
