+++ packages/web3-shared/base/src/specs/index.ts
@@ -11,7 +11,7 @@ import type {
     BindingProof,
     ProfileIdentifier,
     Proof,
-} from '@masknet/base'
+} from '@masknet/shared-base'
 import type { api } from '@dimensiondev/mask-wallet-core/proto'
 import type {
     ReturnChainResolver,
