+++ packages/web3-shared/evm/package.json
@@ -15,7 +15,7 @@
     "@ethersproject/bytes": "5.7.0",
     "@ethersproject/keccak256": "5.7.0",
     "@ethersproject/strings": "5.7.0",
-    "@masknet/base":"workspace:^",
+    "@masknet/shared-base": "workspace:^",
     "@masknet/web3-constants": "workspace:^",
     "@masknet/web3-contracts": "workspace:^",
     "@masknet/web3-shared-base": "workspace:^",
