+++ packages/web3-shared/evm/src/helpers/address.ts
@@ -1,6 +1,6 @@
 import { EthereumAddress } from 'wallet.ts'
 import { getEnumAsArray } from '@masknet/kit'
-import { isPopupPage } from '@masknet/base'
+import { isPopupPage } from '@masknet/shared-base'
 import { isSameAddress } from '@masknet/web3-shared-base'
 import { getENSConstants, getRedPacketConstants, getTokenConstant, ZERO_ADDRESS } from '../constants/index.js'
 import { ChainId, NetworkType, ProviderType } from '../types/index.js'
