+++ packages/web3-shared/evm/src/helpers/formatter.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from 'bignumber.js'
 import { EthereumAddress } from 'wallet.ts'
-import { createLookupTableResolver } from '@masknet/base'
+import { createLookupTableResolver } from '@masknet/shared-base'
 import { isValidDomain } from './isValidDomain.js'
 import { isValidAddress } from './address.js'
 import { SchemaType } from '../types/index.js'
