+++ packages/web3-shared/evm/src/libs/ConnectionContext.ts
@@ -1,6 +1,6 @@
 import type { RequestArguments } from 'web3-core'
+import type { ECKeyIdentifier } from '@masknet/shared-base'
 import type { Connection, ConnectionOptions } from '@masknet/web3-shared-base'
-import type { ECKeyIdentifier } from '@masknet/base'
 import { ErrorEditor } from './ErrorEditor.js'
 import { PayloadEditor } from './PayloadEditor.js'
 import { createJsonRpcPayload, createJsonRpcResponse } from '../helpers/provider.js'
