+++ packages/web3-shared/evm/src/libs/Signer.ts
@@ -1,4 +1,4 @@
-import { SignType } from '@masknet/base'
+import { SignType } from '@masknet/shared-base'
 import type { Transaction } from '../types/index.js'
 
 export class Signer<T> {
