+++ packages/web3-shared/evm/tsconfig.json
@@ -6,5 +6,10 @@
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
   "include": ["./src", "./src/**/*.json"],
-  "references": [{ "path": "../base" }, { "path": "../../injected-script/sdk" }, { "path": "../../polyfills" }]
+  "references": [
+    { "path": "../base" },
+    { "path": "../../shared-base" },
+    { "path": "../../injected-script/sdk" },
+    { "path": "../../polyfills" }
+  ]
 }
