+++ packages/web3-shared/flow/package.json
@@ -12,7 +12,7 @@
   "types": "./dist/index.d.js",
   "dependencies": {
     "@blocto/fcl": "^0.0.78-alpha.10",
-    "@masknet/base":"workspace:^",
+    "@masknet/shared-base": "workspace:^",
     "@masknet/web3-constants": "workspace:^",
     "@masknet/web3-shared-base": "workspace:^",
     "@onflow/types": "^0.0.5",
