+++ packages/web3-shared/flow/src/constants/descriptors.ts
@@ -1,5 +1,5 @@
 import { getEnumAsArray } from '@masknet/kit'
-import { EnhanceableSite, ExtensionSite, NetworkPluginID } from '@masknet/base'
+import { EnhanceableSite, ExtensionSite, NetworkPluginID } from '@masknet/shared-base'
 import {
     type ChainDescriptor,
     createFungibleToken,
