+++ packages/web3-shared/flow/src/helpers/domain.ts
@@ -1,4 +1,4 @@
-import { createLookupTableResolver } from '@masknet/base'
+import { createLookupTableResolver } from '@masknet/shared-base'
 import { SchemaType } from '../types.js'
 
 export function formatDomainName(domain?: string, size?: number) {
