+++ packages/web3-shared/flow/tsconfig.json
@@ -6,5 +6,5 @@
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
   "include": ["./src", "./src/**/*.json"],
-  "references": [{ "path": "../base" }]
+  "references": [{ "path": "../base" }, { "path": "../../shared-base" }]
 }
