+++ packages/web3-shared/solana/src/helpers/address.ts
@@ -2,9 +2,9 @@ import bs58 from 'bs58'
 import * as Web3 from '@solana/web3.js'
 import { getEnumAsArray } from '@masknet/kit'
 import { isSameAddress } from '@masknet/web3-shared-base'
-import { createLookupTableResolver } from '@masknet/base'
 import { ChainId, NetworkType, ProviderType, SchemaType } from '../types.js'
 import { getTokenConstant, ZERO_ADDRESS } from '../constants/index.js'
+import { createLookupTableResolver } from '@masknet/shared-base'
 
 export function encodePublicKey(key: Web3.PublicKey) {
     return key.toBase58()
