+++ packages/web3-shared/solana/tsconfig.json
@@ -6,5 +6,5 @@
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
   "include": ["./src", "./src/**/*.json"],
-  "references": [{ "path": "../../theme" }, { "path": "../../web3-shared/base" }, { "path": "../../base" }]
+  "references": [{ "path": "../../theme" }, { "path": "../../web3-shared/base" }, { "path": "../../web3-shared/evm" }]
 }
