+++ packages/web3-telemetry/src/hooks/useUnmountReport.ts
@@ -1,6 +1,6 @@
 import { useUnmount } from 'react-use'
 import { TelemetryAPI } from '@masknet/web3-providers/types'
-import { type Unresolved, resolve } from '@masknet/base'
+import { type Unresolved, resolve } from '@masknet/shared-base'
 import { useTelemetry } from './useTelemetry.js'
 
 /**
